Pertemuan 4
Pengenalan Array (Integer)

+++ COntohBuiltInFunction_Pow.cpp
@@ -6,9 +6,9 @@
 int main (){
 	int a=2;
 	int b=2;
-	int pow_result;
+	float pow_result;
 	pow_result = pow(a,b);
 	
-	printf("Contoh Built-in Function POW : % d \n", pow_result); // untuk perpangkat
+	printf("Contoh Built-in Function POW : % f \n", pow_result); // untuk perpangkat
 	return 0;
 }
